squads/mmos-squad/requirements.txt
# MMOS Squad - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies (required)
pyyaml>=6.0
requests>=2.28.0

# Optional dependencies - uncomment if needed

# Google Gemini LLM integration (for gemini_analyzer.py)
# google-generativeai>=0.3.0

# Supabase database persistence (for db_persister.py, sources_importer.py)
# supabase>=2.0.0
